Update BNS register/update username for free test to use feePayer
